Правка DuplicateKeyDetector, лоадеров, unsigned -> uint32

+++ SFML1/src/core/ecs/systems/debug_overlay_system.cpp
@@ -125,7 +125,7 @@ namespace core::ecs {
             return;
         }
         mFpsText->setPosition(props.position);
-        mFpsText->setCharacterSize(props.characterSize);
+        mFpsText->setCharacterSize(static_cast<unsigned int>(props.characterSize));
         mFpsText->setFillColor(props.color);
     }
